Mvc refactor

+++ src/utils/index.js
@@ -1,6 +1,6 @@
 import { getCategories } from '../models/category/index.js';
 
-const getNav = async () => {
+const getNav = async (user) => {
     const categories = await getCategories();
     let nav = '<nav><ul>';
     categories.forEach((row) => {
@@ -8,14 +8,29 @@ const getNav = async () => {
         const name = row.category_name;
         nav += `<li><a href="/category/view/${id}">${name}</a></li>`
     });
-    return `
-    ${nav}
+    
+    if (user) {
+        nav += `
         <li><a href="/game/add">Add Game</a></li>
         <li><a href="/category/add">Add Category</a></li>
         <li><a href="/category/delete">Delete Category</a></li>
         <li><a href="/About">About Me</a></li>
-        </ul>
-    </nav>`;
+        <li><a href="/account/logout">Logout</a></li>`;
+    } else {
+        nav += `
+        <li><a href="/account/register">Register</a></li>
+        <li><a href="/account/login">Login</a></li>`;
+    }
+
+    return `${nav}</ul></nav>`;
 };
 
-export { getNav };
+const requireAuth = (req, res, next) => {
+    if (!req.session.user) {
+        req.flash("error", "You must be logged in to access this page.");
+        return res.redirect("/account/login");
+    }
+    next();
+};
+
+export { requireAuth, getNav };
